Fix 2023_10_22_212643_create_categories_table.php

+++ database/migrations/admin/2024_01_17_204100_create_reports_table.php
@@ -29,10 +29,12 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('reports', function (Blueprint $table) {
-            $table->dropForeignIdFor(Category::class);
-        });
+        if (App::isLocal()) {
+            Schema::table('reports', function (Blueprint $table) {
+                $table->dropForeignIdFor(Category::class);
+            });
 
-        Schema::dropIfExists('reports');
+            Schema::dropIfExists('reports');
+        }
     }
 };
